chore: organise files and folders for Map and Fiche components
chore: rebase package-lock.json

chore: rebase with search component
