+++ application/app/components/fiches/ficheEtablissement/ProtectionCard.tsx
