+++ application/package.json
@@ -22,6 +22,7 @@
     "tailwind-merge": "^3.0.2",
     "tailwindcss-animate": "^1.0.7",
     "@turf/turf": "^7.2.0",
+    "lucide-react": "^0.487.0",
     "maplibre-gl": "^5.2.0",
     "next": "15.1.6",
     "react": "^19.0.0",
